requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
numpy>=1.24.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
scipy>=1.11.0
